domain refactoring
