+++ .gitignore
@@ -1,3 +1,3 @@
+__pycache__/
+.vscode/
 .env
-/.venv/
-/pgdata/
