+++ AuthorizationService/Makefile
@@ -1,0 +1,26 @@
+DC = docker-compose
+EXEC = docker exec -it
+LOGS = docker logs
+ENV = --env-file .env
+APP_FILE = docker_compose/app.yaml
+APP_CONTAINER = authorization-service
+
+.PHONY: app
+app:
+	${DC} -f ${APP_FILE} ${ENV} up --build -d
+
+.PHONY: app-up
+app-up:
+	${DC} -f ${APP_FILE} ${ENV} up -d
+
+.PHONY: app-down
+app-down:
+	${DC} -f ${APP_FILE} down
+
+.PHONY: app-shell
+app-shell:
+	${EXEC} ${APP_CONTAINER} bash
+
+.PHONY: app-logs
+app-logs:
+	${LOGS} ${APP_CONTAINER} -f
