+++ AuthorizationService/app/domain/entities/base.py
