+++ AuthorizationService/app/domain/entities/user.py
@@ -4,12 +4,14 @@ from domain.entities.base import BaseEntity
 from domain.values.email import Email
 from domain.values.password import Password
 from domain.values.login import Login
+from domain.values.role import Role
 
 @dataclass
 class User(BaseEntity):
     login: Login
     password: Password
     email: Email
+    role: Role
