+++ AuthorizationService/app/domain/exceptions/base.py
