+++ AuthorizationService/app/domain/values/password.py
@@ -1,9 +1,7 @@
 from dataclasses import dataclass
 
-
 from domain.exceptions.password import PasswordIsEmptyException, PasswordTooLongException
 from domain.values.base import BaseValueObject
-
 
 
 @dataclass(frozen=True)
@@ -11,7 +9,7 @@ class Password(BaseValueObject):
     value: str
     
     def validate(self):
-        if len(self.value) > 255:
+        if len(self.value) > 64:
             raise PasswordTooLongException(self.value)
         if not self.value:
             raise PasswordIsEmptyException(self.value)
