+++ AuthorizationService/docker_compose/app.yaml
@@ -1,0 +1,14 @@
+services:
+  authorization-service:
+    container_name: authorization-service
+    build: 
+      context: ..
+      dockerfile: Dockerfile
+    command: "uvicorn --factory application.api.main:create_app --host=0.0.0.0 --reload --port 8000"
+    volumes:
+      - ../app:/app/
+    ports:
+      - "${API_PORT}:8000"
+    env_file:
+      - ../.env
+    restart: always
