--- AuthorizationService/dockerfile
@@ -1,14 +1,0 @@
-FROM python:3.13-slim
-
-ENV PYTHONDONTWRITEBYTECODE 1
-ENV PYTHONUNBUFFERED 1
-ENV PIP_ROOT_USER_ACTION=ignore
-
-# Adding project files in container   
-RUN mkdir -p /app
-COPY ./AuthorizationService /app
-
-# Builing a project
-WORKDIR /app
-
-RUN pip install --no-cache-dir -r requirements.txt
