+++ AuthorizationService/pyproject.toml
@@ -1,0 +1,17 @@
+[tool.poetry]
+name = "authorizationservice"
+version = "0.1.0"
+description = ""
+authors = ["JonAJ <jonajmail@gmail.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.12"
+fastapi = "^0.115.5"
+uvicorn = {extras = ["all"], version = "^0.32.0"}
+ipython = "^8.29.0"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
