--- docker-compose.yml
@@ -1,33 +1,0 @@
-services:
-  
-  authorization-service:
-    container_name: authorization_service
-    build: 
-      context: .
-      dockerfile: ./AuthorizationService/dockerfile
-    depends_on:
-      - database
-    command: uvicorn --factory application.api.main:create_app --host=0.0.0.0 --reload
-    volumes:
-      - ./AuthorizationService:/app
-    ports:
-      - ${APP_HOST_PORT}:8000
-    restart: always
-    env_file:
-      - .env
-  
-  database:
-    container_name: postgres
-    image: postgres:15.2
-    environment:
-      - POSTGRES_DB=postgres
-      - POSTGRES_USER=${POSTGRESS_DB_USER}
-      - POSTGRES_PASSWORD=${POSTGRESS_DB_PASSWORD}
-    ports:
-      - ${POSTGRESS_DB_HOST_PORT}:5432
-    volumes:
-      - ./pgdata:/var/lib/posgresql/data
-    env_file:
-      - .env
-    restart: always
-
